+++ webAdopcion/src/main/java/fluffandpaws/webadopcion/Repositories/AnimalRepository.java
@@ -6,12 +6,13 @@ package fluffandpaws.webadopcion.Repositories;
 -----------------------------------------------------------*/
 
 import fluffandpaws.webadopcion.BBDD.Animales;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 //import org.springframework.data.repository.CrudRepository;
 @Repository
-public interface AnimalRepository extends CrudRepository<Animales, Integer> {
+public interface AnimalRepository extends CrudRepository<Animales, Long> {
     //Igual que el repositorio de usuarios
 }
