+++ webAdopcion/src/main/java/fluffandpaws/webadopcion/Repositories/UserRepository.java
@@ -1,10 +1,11 @@
 package fluffandpaws.webadopcion.Repositories;
 
 import fluffandpaws.webadopcion.BBDD.Usuarios;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 //Spring lo implementa automáticamente
 @Repository
-public interface UserRepository extends CrudRepository<Usuarios, Integer> {
+public interface UserRepository extends CrudRepository<Usuarios, Long> {
     //CRUD = Create, Read, Update, Delete
 }
